[fe] update dev
