+++ src/core/CMakeLists.txt
@@ -11,7 +11,7 @@ macro(CORE_ADD_SOURCE)
     set(CoreSourceFiles ${CoreSourceFiles} ${SOURCE_FILES} PARENT_SCOPE)
 endmacro(CORE_ADD_SOURCE)
 
-set(CoreSourceFiles "${SparrowCoreDir}/SparrowCore.h" IO/AttriVertex.h Model/Mesh.h Model/Model.h Model/Vertex.h)
+set(CoreSourceFiles "${SparrowCoreDir}/SparrowCore.h")
 
 # --------------------------- PLEASE ADD SUBMODULES HERE -------------------
 add_subdirectory(Utils)
