+++ src/core/Model/Model.h
@@ -78,9 +78,7 @@ namespace SPW
             m_Meshes = std::move(_meshes);
 	    }
 
-    	~Model() {
-            std::cout << " ";
-        };
+    	~Model() {}
 
         [[nodiscard]] const std::vector<std::shared_ptr<Mesh>>& GetMeshes() const { return m_Meshes; }
         void AddMesh(std::shared_ptr<Mesh> _mesh) { m_Meshes.emplace_back(std::move(_mesh)); }
